requirements.txt
numpy==1.19.5
opencv-contrib-python-headless==4.7.0.68
opencv-python==4.7.0.72
opencv-python-headless==4.5.4.60
pycocotools==2.0.6
jsonlines==3.1.0
jsonschema 
matplotlib 
matplotlib-inline
Pillow==9.2.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm
trimesh==3.23.5
openpifpaf==0.13.11
detectron2==0.6